[manifest] Roll zircon 5b0190f..72667b6 (2 commits)
72667b6 [fidl] Simplify the enum type representation
56aff58 [debugger] Clarify thread state API

+++ lib/inferior_control/registers.cc
@@ -45,7 +45,7 @@ bool Registers::RefreshRegsetHelper(int regset, void* buf, size_t buf_size) {
     return true;
   }
 
-  uint32_t regset_size;
+  size_t regset_size;
   zx_status_t status = zx_thread_read_state(
     thread()->handle(), regset, buf, buf_size, &regset_size);
   if (status < 0) {
